requirements.txt
aiogram==2.25.2
comtypes==1.1.14
python-pptx==0.6.23
openpyxl==3.1.2
Pillow==10.3.0
telethon==1.34.0
python-dotenv==1.0.0